Fix broken build due to undeclared dependency on cartesian-product-switch.
While doing my big refactoring I removed an instance of cartesian-product-switch which I thought was the only one in the codebase. But even if that had been true, I forgot to remove the dependency in package.lisp.

I guess it's only fair that I should make one big stupid mistake when doing a reckless night of intense coding while tired.